refactor(context): organize compaction as package
