+++ nemo_gym/context_compaction/__init__.py
@@ -1,0 +1,127 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Semantic context compaction for multi-turn Gym rollouts."""
+
+from nemo_gym.context_compaction.config import (
+    CompactionScheduleConfig,
+    ContextGuardConfig,
+    ContextHistoryConfig,
+    HistoryPolicyConfig,
+    ImageRecencyConfig,
+    ReasoningRecencyConfig,
+    RecencyHistoryPolicyConfig,
+)
+from nemo_gym.context_compaction.controller import (
+    HistoryController,
+    TurnChunkedHistoryController,
+    build_guard_outcome_records,
+    evaluate_context_guards,
+    pending_observation_group_ids,
+)
+from nemo_gym.context_compaction.history import (
+    CanonicalHistory,
+    ContextMeasurements,
+    FinalizedChunkRecord,
+    GenerationContract,
+    GuardOutcomeRecord,
+    MaterializedHistoryView,
+    MediaOccurrence,
+    ObservedCompletion,
+    PolicyDecisionEvidence,
+    PolicyDecisionRecord,
+    PolicyOutputSpan,
+    PreparedHistoryView,
+    RewriteBoundaryEvent,
+    SemanticHistory,
+    TransformationLineageDeltaRecord,
+    UnitLineageRecord,
+    build_lineage_delta,
+    canonical_digest,
+    capture_observed_completion,
+    normalize_semantic_items,
+    register_semantic_part_kind,
+    stable_id,
+    unregister_semantic_part_kind,
+)
+from nemo_gym.context_compaction.materialization import (
+    descriptor_is_append_compatible,
+    materialize_history_view,
+    ordered_media_is_append_compatible,
+)
+from nemo_gym.context_compaction.policies import (
+    HistoryPolicy,
+    IdentityHistoryPolicy,
+    RecencyHistoryPolicy,
+    build_history_policy,
+    register_history_policy,
+    unregister_history_policy,
+)
+from nemo_gym.context_compaction.session import (
+    ContextCompactedResponse,
+    ContextCompactedTransportResponse,
+    ContextCompactionContract,
+    ContextCompactionSession,
+    ModelCallMetadata,
+    PreparedContextCompactionCall,
+    TransportContextCompactionContract,
+    build_generation_contract,
+    build_transport_response,
+)
+
+
+__all__ = [
+    "CanonicalHistory",
+    "CompactionScheduleConfig",
+    "ContextCompactedResponse",
+    "ContextCompactedTransportResponse",
+    "ContextCompactionContract",
+    "ContextCompactionSession",
+    "ContextGuardConfig",
+    "ContextHistoryConfig",
+    "ContextMeasurements",
+    "FinalizedChunkRecord",
+    "GenerationContract",
+    "GuardOutcomeRecord",
+    "HistoryController",
+    "HistoryPolicy",
+    "HistoryPolicyConfig",
+    "IdentityHistoryPolicy",
+    "ImageRecencyConfig",
+    "MaterializedHistoryView",
+    "MediaOccurrence",
+    "ModelCallMetadata",
+    "ObservedCompletion",
+    "PolicyDecisionEvidence",
+    "PolicyDecisionRecord",
+    "PolicyOutputSpan",
+    "PreparedContextCompactionCall",
+    "PreparedHistoryView",
+    "ReasoningRecencyConfig",
+    "RecencyHistoryPolicy",
+    "RecencyHistoryPolicyConfig",
+    "RewriteBoundaryEvent",
+    "SemanticHistory",
+    "TransformationLineageDeltaRecord",
+    "TransportContextCompactionContract",
+    "TurnChunkedHistoryController",
+    "UnitLineageRecord",
+    "build_generation_contract",
+    "build_guard_outcome_records",
+    "build_history_policy",
+    "build_lineage_delta",
+    "build_transport_response",
+    "canonical_digest",
+    "capture_observed_completion",
+    "descriptor_is_append_compatible",
+    "evaluate_context_guards",
+    "materialize_history_view",
+    "normalize_semantic_items",
+    "ordered_media_is_append_compatible",
+    "pending_observation_group_ids",
+    "register_history_policy",
+    "register_semantic_part_kind",
+    "stable_id",
+    "unregister_history_policy",
+    "unregister_semantic_part_kind",
+]
